+++ src/main/java/org/example/goormssd/usermanagementbackend/security/CustomUserDetailsService.java
@@ -1,4 +1,4 @@
-package org.example.goormssd.usermanagementbackend.config.security;
+package org.example.goormssd.usermanagementbackend.security;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
